requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
sse-starlette
python-multipart
PyPDF2
python-docx